requirements.txt
Flask==2.0.1
python-telegram-bot>=19.6
pyTelegramBotAPI==3.8.2
requests==2.26.0
user-agents==2.2.0
python-dotenv==0.19.2
gunicorn==20.1.0
Werkzeug==2.0.0